refactor: change RegisterPanelDetails to RegisterFormData in schemas and types
